Extract remaining items and cleanup
